Установил spring.jpa.hibernate.ddl-auto=validate Пошли нестыковки. Изменил ключи на bigint, до этого стоял int8 (serial) Валидация проходит теперь

+++ src/main/java/com/nursery/nursery_api/model/Report.java
@@ -34,7 +34,7 @@ public class Report {
     private Long forteit;
 
     @Column(name = "day_report")
-    private Long dayReport;
+    private Integer dayReport;
 
     @JsonIgnore
     @OneToMany(mappedBy = "report",fetch = FetchType.EAGER)
